[FLINK-1371] [runtime] Fix KryoSerializer to not swallow EOFExceptions
