+++ flink-java/src/main/java/org/apache/flink/api/java/typeutils/runtime/KryoSerializer.java
@@ -30,6 +30,7 @@ import org.apache.flink.core.memory.DataOutputView;
 
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
+import java.io.EOFException;
 import java.io.IOException;
 
 public class KryoSerializer<T> extends TypeSerializer<T> {
@@ -112,8 +113,19 @@ public class KryoSerializer<T> extends TypeSerializer<T> {
 			previousOut = target;
 		}
 		
-		kryo.writeClassAndObject(output, record);
-		output.flush();
+		try {
+			kryo.writeClassAndObject(output, record);
+			output.flush();
+		}
+		catch (KryoException ke) {
+			Throwable cause = ke.getCause();
+			if (cause instanceof EOFException) {
+				throw (EOFException) cause;
+			}
+			else {
+				throw ke;
+			}
+		}
 	}
 
 	@Override
@@ -169,5 +181,14 @@ public class KryoSerializer<T> extends TypeSerializer<T> {
 			this.kryo.register(type);
 			this.kryo.setClassLoader(Thread.currentThread().getContextClassLoader());
 		}
+	}
+	
+	// --------------------------------------------------------------------------------------------
+	// for testing
+	// --------------------------------------------------------------------------------------------
+	
+	Kryo getKryo() {
+		checkKryoInitialized();
+		return this.kryo;
 	}
 }
